Fix bug in TupleCutter.cc
When testing for Electron 0 in EB and Electron 1 in EE, Electron 0 was
mistakenly tested both times, and hence always failed because it is
impossible for it to be in both locations at the same time. This caused
events with Electron 0 in EB to always be rejected, and hence removed
about half of the events.