Add in Background Toy and adjust fits
bg_func.nb is a Mathematic 8 notebook that lets you adjust the
parameters of the fit function and see the plotted result.

Using this I've adjust the limits of the fitting function to more fully
cover the acceptable range of fits.

I've also written down roughly what each parameter does in
BackgroundFunctions.h.