Updates to fitEffAcc and removal of BackgroundFitter
We no longer use BackgroundFitter for the trigger, and it is unlikely we
will use it in the future, so it has been removed.

We have updated BackgroundFunctions.cc to allow us to fit alpha, beta,
gamma, and delta instead of fixing them before hand.

fitEffAcc now fits the background first, and then fits the signal. It no
longer takes in background shapes. Also it no longer expects normalized
histograms, and so there have been some changes in the efficiency
calculation to account for this.