Start ZSmearer
Add in ZSmear to take in a ZEffTree::ZInfo, apply a smearing, and return
the modified ZInfo.

A TRandom* is passed into the function, so that the user can decide what
TRandom object they want, and also so that there are no issues with
creating and destroying a new generator each iteration of a loop.